+++ internal/provider/provider.go
@@ -20,7 +20,6 @@ import (
 	authorizedentraapps "github.com/axiansinfoma/terraform-provider-bcadmincenter/internal/services/authorized_entra_apps"
 	"github.com/axiansinfoma/terraform-provider-bcadmincenter/internal/services/available_applications"
 	environmentapps "github.com/axiansinfoma/terraform-provider-bcadmincenter/internal/services/environment_apps"
-	environmentsettings "github.com/axiansinfoma/terraform-provider-bcadmincenter/internal/services/environment_settings"
 	environmentsupportcontact "github.com/axiansinfoma/terraform-provider-bcadmincenter/internal/services/environment_support_contact"
 	"github.com/axiansinfoma/terraform-provider-bcadmincenter/internal/services/environments"
 	notificationrecipients "github.com/axiansinfoma/terraform-provider-bcadmincenter/internal/services/notification_recipients"
@@ -165,7 +164,6 @@ func (p *BCAdminCenterProvider) Resources(ctx context.Context) []func() resource
 		environments.NewEnvironmentResource,
 		environments.NewUpdateScheduleResource,
 		environmentapps.NewEnvironmentAppResource,
-		environmentsettings.NewEnvironmentSettingsResource,
 		environmentsupportcontact.NewEnvironmentSupportContactResource,
 		notificationrecipients.NewNotificationRecipientResource,
 	}